+++ src/elements/autocomplete/autocomplete-base-element.ts
@@ -185,7 +185,8 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
   /** Opens the autocomplete. */
   public open(): void {
     if (
-      this.state !== 'closed' ||
+      this.state === 'opening' ||
+      this.state === 'opened' ||
       !this._overlay ||
       this.options.length === 0 ||
       this._readonly() ||
@@ -205,6 +206,8 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
     }
     this._setOverlayPosition(originElement);
     this._setNextActiveOptionIfAutoActiveFirstOption();
+    this._attachOpenPanelEvents();
+    this._escapableOverlayController.connect();
 
     // If the animation duration is zero, the animationend event is not always fired reliably.
     // In this case we directly set the `opened` state.
@@ -215,7 +218,7 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
 
   /** Closes the autocomplete. */
   public close(): void {
-    if (this.state !== 'opened' || !this.dispatchBeforeCloseEvent()) {
+    if (this.state === 'closing' || this.state === 'closed' || !this.dispatchBeforeCloseEvent()) {
       return;
     }
 
@@ -476,12 +479,10 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
 
   private _handleOpening(): void {
     this.state = 'opened';
-    this._attachOpenPanelEvents();
     if (this.originElement) {
       this._originResizeObserver.observe(this.originElement);
     }
     this.triggerElement?.setAttribute('aria-expanded', 'true');
-    this._escapableOverlayController.connect();
     this.dispatchOpenEvent();
   }
 
@@ -489,6 +490,7 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
     this.state = 'closed';
     this.hidePopover?.();
     this.triggerElement?.setAttribute('aria-expanded', 'false');
+
     // Clears the input if there's user interaction without selection (selection clears `_lastUserInput`).
     if (this.requireSelection && this.triggerElement && this._lastUserInput) {
       const setValue = Object.getOwnPropertyDescriptor(HTMLInputElement.prototype, 'value')!.set!;
@@ -551,6 +553,20 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
         capture: true,
       },
     );
+
+    this.triggerElement?.addEventListener(
+      'blur',
+      (e) => {
+        // If the new focus is the autocomplete or inside of it then an option
+        // was selected and there is a separate mechanism that closes this instance.
+        if (!this.contains(e.relatedTarget as Node)) {
+          this.close();
+        }
+      },
+      {
+        signal: this._openPanelEventsController.signal,
+      },
+    );
   }
 
   // Check if the pointerdown event target is triggered on the menu.
@@ -570,12 +586,11 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
   };
 
   private _closedPanelKeyboardInteraction(event: KeyboardEvent): void {
-    if (this.state !== 'closed') {
+    if (this.state === 'opening' || this.state === 'opened') {
       return;
     }
 
     switch (event.key) {
-      case 'Enter':
       case 'ArrowDown':
       case 'ArrowUp':
         this.open();
@@ -592,6 +607,9 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
   }
 
   protected override render(): TemplateResult {
+    // Scroll areas without containing an interactive element will receive focus when tabbing through the document.
+    // If there are a lot of options and when pressing tab key, the scroll area on sbb-autocomplete__options gets focus.
+    // As elements inside the panel should never get focus, we have to avoid that by setting tabindex=-1.
     return html`
       <div class="sbb-autocomplete__gap-fix"></div>
       <div class="sbb-autocomplete__container">
@@ -606,6 +624,7 @@ export abstract class SbbAutocompleteBaseElement<T = string> extends SbbNegative
               class="sbb-autocomplete__options"
               role=${!ariaRoleOnHost ? this.panelRole : nothing}
               id=${!ariaRoleOnHost ? this.overlayId : nothing}
+              tabindex="-1"
               ${ref((containerRef) => (this._optionContainer = containerRef as HTMLElement))}
             >
               <slot @slotchange=${this._handleSlotchange}></slot>